+++ app/Modules/Trader/Indicators/EMA.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Modules\Trader\Indicators;
 
 final class EMA extends AbstractIndicator