+++ docker/8.3/stubs/trader/functions/trader-cdlrisefall3methods.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * Rising/Falling Three Methods
  *