+++ docker/8.3/stubs/trader/functions/trader-cdlshootingstar.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * Shooting Star
  *