+++ docker/8.3/stubs/trader/functions/trader-minindex.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * Index of lowest value over a specified period
  *