+++ tests/Integration/Services/BrokerClientServiceTest.php
@@ -1,25 +1,53 @@
 <?php
 
-declare(strict_types=1);
-
 use App\Modules\Trader\Configs\BrokerClientConfig;
+use App\Modules\Trader\Payloads\LoginPayload;
 use App\Modules\Trader\Services\BrokerClientService;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Str;
 use WebSocket\Client;
 
-it('initializes websocket client', function () {
-    $config = new class('ws://example.com', 'userId', 'password') extends BrokerClientConfig
-    {
-    };
+it('initializes, authenticates the client correctly', function () {
+    Config::set([
+        'broker.uri' => 'ws://fake-uri',
+        'broker.user_id' => 'fake-user-id',
+        'broker.password' => 'fake-password',
+    ]);
 
-    $client = new class($config) extends BrokerClientService
+    $loginPayload = new LoginPayload(config('broker.user_id'), config('broker.password'));
+
+    $clientMock = Mockery::mock(Client::class);
+
+    $clientMock->shouldReceive('isConnected')
+        ->andReturn(false);
+
+    $clientMock->shouldReceive('send')
+        ->once()
+        ->with($loginPayload->toJson());
+
+    $clientMock->shouldReceive('receive')
+        ->once()
+        ->andReturn(json_encode([
+            'status' => true,
+            'streamSessionId' => Str::random(32),
+        ]));
+
+    $this->app->instance(Client::class, $clientMock);
+
+    $brokerClientConfig = $this->app->make(BrokerClientConfig::class);
+
+    $brokerClientService = new class($brokerClientConfig, $clientMock) extends BrokerClientService
     {
-        public function getClient(): Client
+        public function __construct(BrokerClientConfig $config, Client $client)
         {
-            return $this->client;
+            parent::__construct($config);
+
+            $this->client = $client;
         }
     };
 
-    $socketClient = $client->getClient();
+    $response = $brokerClientService->getClient();
 
-    expect($socketClient)->toBeInstanceOf(Client::class);
+    expect($response)->toBeInstanceOf(Client::class)
+        ->and($response->isConnected())->toBeFalse();
 });